Switch C-p mode to MRU. Swap RainbowParentheses for Rainbow.
Rainbow Parens requires options that error if not installed
correctly and even then It is hit and miss if it will work.
Rainbow just works and also highlight operators inside the brackets which
is going to come in handy.

With the addition of bufexplorer 90% of what C-p is doing now is MRU
work. If I can find a way that allows to fuzz from ~/code whenever
I open it that'll be much better.